requirements.txt
streamlit
google-play-scraper
groq
google-generativeai
python-dotenv
pandas
langdetect
